updates for review comments
