+++ hsds/attr_sn.py
@@ -27,8 +27,9 @@ from .util.domainUtil import getBucketForDomain, verifyRoot
 from .util.attrUtil import validateAttributeName, getRequestCollectionName
 from .util.hdf5dtype import validateTypeItem, getBaseTypeJson
 from .util.hdf5dtype import createDataType, getItemSize
-from .util.arrayUtil import jsonToArray, getShapeDims, getNumElements
+from .util.arrayUtil import jsonToArray, getNumElements
 from .util.arrayUtil import bytesArrayToList
+from .util.dsetUtil import getShapeDims
 from .servicenode_lib import getDomainJson, getObjectJson, validateAction
 from . import hsds_logger as log
 from . import config
